listas

+++ Aulas/03-input.py
@@ -1,7 +1,7 @@
 faturamento = input("digite o valor do faturamento desse mes: ") # input por padrao eh uma string
 faturamento = faturamento.replace("mzn", " ") #susbtitui o mzn por um vazio
 
-faturamento = int(faturamento) # converte string do input pra inteiro
+faturamento = float(faturamento) # converte string do input pra inteiro
 print(f"O faturamento desse mes foi de {faturamento}")
 
 custo = 300
